qc: move 'QCOutputs' class from 'reporting' to 'outputs' submodule.
