+++ auto_process_ngs/qc/reporting.py
@@ -47,7 +47,7 @@ from .cellranger import CellrangerMulti
 from .outputs import fastqc_output
 from .outputs import fastq_screen_output
 from .outputs import fastq_strand_output
-from .outputs import expected_outputs
+from .outputs import QCOutputs
 from .plots import useqlenplot
 from .plots import ureadcountplot
 from .plots import uscreenplot
@@ -545,928 +545,6 @@ class QCProject:
         self.output_files = qc_outputs.output_files
         # Sequence length stats
         self.stats = AttributeDictionary(**qc_outputs.stats)
-
-class QCOutputs:
-    """
-    Class to detect and characterise QC outputs
-
-    On instantiation this class scans the supplied
-    directory to identify and classify various
-    artefacts which are typically produced by the
-    QC pipeline.
-
-    The following attributes are available:
-
-    - fastqs: sorted list of Fastq names
-    - reads: list of reads (e.g. 'r1', 'r2', 'i1' etc)
-    - samples: sorted list of sample names extracted
-      from Fastqs
-    - fastq_screens: sorted list of screen names
-    - cellranger_references: sorted list of reference
-      datasets used with 10x pipelines
-    - multiplexed_samples: sorted list of sample names
-      for multiplexed samples (e.g. 10x CellPlex)
-    - outputs: list of QC output categories detected (see
-      below for valid values)
-    - output_files: list of absolute paths to QC output
-      files
-    - software: dictionary with information on the
-      QC software packages
-    - stats: AttrtibuteDictionary with useful stats from
-      across the project
-    - config_files: list of QC configuration files found
-      in the QC directory (see below for valid values)
-
-    The following are valid values for the 'outputs'
-    property:
-
-    - 'fastqc_[r1...]'
-    - 'screens_[r1...]'
-    - 'strandedness'
-    - 'sequence_lengths'
-    - 'icell8_stats'
-    - 'icell8_report'
-    - 'cellranger_count'
-    - 'cellranger_multi'
-    - 'cellranger-atac_count'
-    - 'cellranger-arc_count'
-    - 'multiqc'
-
-    The following are valid values for the 'config_files'
-    property:
-
-    - fastq_strand.conf
-    - 10x_multi_config.csv
-    - libraries.<SAMPLE>.csv
-
-    Arguments:
-      qc_dir (str): path to directory to examine
-      fastq_attrs (BaseFastqAttrs): (optional) class for
-        extracting data from Fastq names
-    """
-    def __init__(self,qc_dir,fastq_attrs=None):
-        # Directory to examine
-        self.qc_dir = os.path.abspath(qc_dir)
-        # How to handle Fastq names
-        if fastq_attrs is not None:
-            self.fastq_attrs = fastq_attrs
-        else:
-            self.fastq_attrs = AnalysisFastq
-        # Properties
-        self.fastqs = set()
-        self.samples = []
-        self.reads = []
-        self.stats = AttributeDictionary(
-            max_seqs=None,
-            min_sequence_length_read={},
-            max_sequence_length_read={},
-            min_sequence_length=None,
-            max_sequence_length=None,
-        )
-        self.software = {}
-        self.outputs = set()
-        self.output_files = []
-        self.config_files = []
-        # Raw QC output data
-        self._qc_data = {}
-        # Scan the directory for QC outputs
-        self._collect_qc_outputs()
-        # Report
-        print("Collected %d output files" % len(self.output_files))
-        print("Outputs:")
-        for output in self.outputs:
-            print("- %s" % output)
-        print("Fastqs:")
-        for fq in self.fastqs:
-            print("- %s" % fq)
-        print("Samples:")
-        for s in self.samples:
-            print("- %s" % s)
-        print("Reads:")
-        for r in self.reads:
-            print("- %s" % r)
-        print("Software:")
-        for sw in self.software:
-            print("- %s: %s" % (sw,','.join(self.software[sw])))
-        print("Screens:")
-        for scrn in self.fastq_screens:
-            print("- %s" % scrn)
-        print("QC config files:")
-        for cfg in self.config_files:
-            print("- %s" % cfg)
-
-    def data(self,name):
-        """
-        Return the 'raw' data associated with a QC output
-
-        Arguments:
-          name (str): name identifier for a QC output
-            (e.g. 'fastqc')
-
-        Returns:
-          AttributeDictionary containing the raw data
-            associated with the named QC output.
-
-        Raises:
-          KeyError: if the name doesn't match a stored
-            QC output.
-        """
-        return self._qc_data[name]
-
-    def _collect_qc_outputs(self):
-        """
-        Internal: determine which QC outputs are present
-        """
-        # Get list of files
-        print("Scanning contents of %s" % self.qc_dir)
-        files = [os.path.join(self.qc_dir,f)
-                 for f in os.listdir(self.qc_dir)]
-        print("\t- %d objects found" % len(files))
-        logger.debug("files: %s" % files)
-        # Collect QC outputs
-        for qc_data in (
-                self._collect_fastq_screens(files),
-                self._collect_fastqc(files),
-                self._collect_fastq_strand(files),
-                self._collect_seq_lengths(files),
-                self._collect_icell8(self.qc_dir),
-                self._collect_cellranger_count(self.qc_dir),
-                self._collect_cellranger_multi(self.qc_dir),
-                self._collect_multiqc(self.qc_dir),):
-            self._add_qc_outputs(qc_data)
-        # Collect QC config files
-        self.config_files = self._collect_qc_config_files(files)
-        # Fastq screens
-        fastq_screen = self.data('fastq_screen')
-        self.fastq_screens = sorted(fastq_screen.screen_names)
-        # Sequence lengths
-        seq_lengths = self.data('sequence_lengths')
-        if "sequence_lengths" in seq_lengths.tags:
-            # Maximum number of sequences
-            self.stats['max_seqs'] = seq_lengths.max_seqs
-            # Maximum and minimum sequence lengths for each read
-            for read in seq_lengths.reads:
-                self.stats.min_sequence_length_read[read] = \
-                    seq_lengths.min_seq_length[read]
-                self.stats.max_sequence_length_read[read] = \
-                    seq_lengths.max_seq_length[read]
-            # Maximum and minimum sequence lengths for all reads
-            min_lens = [seq_lengths.min_seq_length[r]
-                        for r in seq_lengths.min_seq_length]
-            if min_lens:
-                self.stats['min_sequence_length'] = min(min_lens)
-            max_lens = [seq_lengths.max_seq_length[r]
-                        for r in seq_lengths.max_seq_length]
-            if max_lens:
-                self.stats['max_sequence_length'] = max(max_lens)
-        # Cellranger count
-        cellranger_count = self.data('cellranger_count')
-        # Single library analyses reference data
-        self.cellranger_references = [ref for ref in
-                                      cellranger_count.references]
-        # Cellranger multi
-        cellranger_multi = self.data('cellranger_multi')
-        # Multiplexed samples
-        self.multiplexed_samples = cellranger_multi.multiplexed_samples
-        for reference_data in cellranger_multi.references:
-            if reference_data not in self.cellranger_references:
-                self.cellranger_references.append(reference_data)
-        self.cellranger_references = sorted(self.cellranger_references)
-        # Fastqs
-        self.fastqs = sorted(list(self.fastqs))
-        # Samples
-        samples = set([self.fastq_attrs(fq).sample_name
-                       for fq in self.fastqs])
-        for s in cellranger_count.samples:
-            samples.add(s)
-        self.samples = sorted(list(samples),
-                              key=lambda s: split_sample_name(s))
-        # Determine reads
-        reads = set()
-        for fastq in self.fastqs:
-            fq = self.fastq_attrs(fastq)
-            if fq.is_index_read:
-                reads.add("i%s" % (fq.read_number
-                                   if fq.read_number is not None else '1'))
-            else:
-                reads.add("r%s" % (fq.read_number
-                                   if fq.read_number is not None else '1'))
-        self.reads = sorted(list(reads))
-        # Sort outputs
-        self.outputs = sorted(self.outputs)
-        # Sort output files
-        self.output_files = sorted(self.output_files)
-
-    def _add_qc_outputs(self,data):
-        """
-        Internal: store 'raw' data for a specific QC output
-
-        The data will be an AttributeDictionary instance
-        returned from one of the ``_collect_...`` methods,
-        which should contain at minimum the following
-        attributes:
-
-        - name: identifier for the QC output (e.g. 'fastqc')
-        - fastqs: list of Fastq names associated with the
-          output
-        - software: a dictionary compromising the names of
-          associated software packages as keys, and a list
-          of versions as the values
-        - output_files: a list of associated output files
-        - tags: a list of output 'classes' associated with
-          the QC output (e.g. 'fastqc_r1')
-
-        The data can be accessed using the name identifier,
-        via the 'data' method; the information about Fastqs,
-        software, output files and tags are added to the
-        global properties.
-        """
-        print("- adding data for '%s'" % data.name)
-        if data.name in self._qc_data:
-            raise KeyError("'%s': data already stored against "
-                           "this name")
-        ##print(data)
-        # Store raw data
-        self._qc_data[data.name] = data
-        # Output files
-        for f in data.output_files:
-            self.output_files.append(f)
-        # Fastqs
-        try:
-            for fq in data.fastqs:
-                self.fastqs.add(fq)
-        except AttributeError:
-            pass
-        # Software versions
-        for sw in data.software.keys():
-            self.software[sw] = [str(v) if v else '?'
-                                 for v in data.software[sw]]
-        # Tags
-        for tag in data.tags:
-            self.outputs.add(tag)
-
-    def _collect_fastq_screens(self,files):
-        """
-        Collect information on FastqScreen outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'fastq_screen'
-        - software: dictionary of software and versions
-        - screen_names: list of associated panel names
-        - fastqs: list of associated Fastq names
-        - fastqs_for_screen: dictionary of panel names
-          and lists of Fastq names associated with each
-          panel
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          files (list): list of file names to examine.
-        """
-        versions = set()
-        output_files = list()
-        fastqs = set()
-        fastqs_for_screen = dict()
-        screen_names = set()
-        tags = set()
-        # Look for legacy screen files
-        legacy_screens = list(filter(lambda f:
-                                     f.endswith("_screen.txt") or
-                                     f.endswith("_screen.png"),
-                                     files))
-        logger.debug("Screens (legacy): %s" % legacy_screens)
-        # Look for new-style screen files
-        screens = list(filter(lambda f:
-                              "_screen_" in os.path.basename(f) and
-                              (f.endswith(".txt") or
-                               f.endswith(".png")),
-                               files))
-        logger.debug("Screens: %s" % screens)
-        print("\t- %d fastq_screen files" % (len(legacy_screens) +
-                                             len(screens)))
-        if legacy_screens:
-            for screen_name in FASTQ_SCREENS:
-                # Explicitly check for legacy screens
-                for screen in list(filter(lambda s:
-                                          s.endswith("_%s_screen.txt" %
-                                                     screen_name),
-                                          legacy_screens)):
-                    fq = self.fastq_attrs(screen[:-len("_screen.txt")])
-                    fastq_name = str(fq)[:-len("_%s" % screen_name)]
-                    tags.add("screens_%s%s" %
-                             (('i' if fq.is_index_read else 'r'),
-                              (fq.read_number
-                               if fq.read_number is not None else '1')))
-                    # Store general information
-                    fastqs.add(fastq_name)
-                    screen_names.add(screen_name)
-                    versions.add(Fastqscreen(screen).version)
-                    # Store Fastq against screen name
-                    if screen_name not in fastqs_for_screen:
-                        fastqs_for_screen[screen_name] = set()
-                    fastqs_for_screen[screen_name].add(fastq_name)
-            # Store the legacy screen files
-            output_files.extend(legacy_screens)
-        if screens:
-            # Pull out the Fastq names from the .txt files
-            for screen in list(filter(lambda s: s.endswith(".txt"),
-                                      screens)):
-                # Assume that names are 'FASTQ_screen_SCREENNAME.txt'
-                fastq_name,screen_name = os.path.basename(screen)\
-                                         [:-len(".txt")].\
-                                         split("_screen_")
-                fq = self.fastq_attrs(fastq_name)
-                tags.add("screens_%s%s" %
-                         (('i' if fq.is_index_read else 'r'),
-                          (fq.read_number
-                           if fq.read_number is not None else '1')))
-                # Store general information
-                fastqs.add(fastq_name)
-                screen_names.add(screen_name)
-                versions.add(Fastqscreen(screen).version)
-                # Store Fastq against screen name
-                if screen_name not in fastqs_for_screen:
-                    fastqs_for_screen[screen_name] = set()
-                fastqs_for_screen[screen_name].add(fastq_name)
-            # Store the screen files
-            output_files.extend(screens)
-        # Return collected information
-        if versions:
-            software = { 'fastq_screen': sorted(list(versions)) }
-        else:
-            software = {}
-        return AttributeDictionary(
-            name='fastq_screen',
-            software=software,
-            screen_names=sorted(list(screen_names)),
-            fastqs=sorted(list(fastqs)),
-            fastqs_for_screen={ s: sorted(list(fastqs_for_screen[s]))
-                                for s in fastqs_for_screen },
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_fastqc(self,files):
-        """
-        Collect information on FastQC outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'fastqc'
-        - software: dictionary of software and versions
-        - fastqs: list of associated Fastq names
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          files (list): list of file names to examine.
-        """
-        versions = set()
-        output_files = list()
-        fastqs = set()
-        tags = set()
-        # Look for fastqc outputs
-        fastqcs = list(
-            filter(lambda f:
-                   f.endswith("_fastqc") and
-                   os.path.exists("%s.html" % f) and
-                   os.path.exists(os.path.join(f,"summary.txt")) and
-                   os.path.exists(os.path.join(f,"fastqc_data.txt")),
-                   files))
-        logger.debug("Fastqc: %s" % fastqcs)
-        print("\t- %d fastqc files" % len(fastqcs))
-        if fastqcs:
-            versions = set()
-            # Pull out the Fastq names from the Fastqc files
-            for fastqc in fastqcs:
-                f = os.path.basename(fastqc)[:-len("_fastqc")]
-                fastqs.add(f)
-                fq = self.fastq_attrs(f)
-                tags.add("fastqc_%s%s" %
-                         (('i' if fq.is_index_read else 'r'),
-                          (fq.read_number
-                           if fq.read_number is not None else '1')))
-                versions.add(Fastqc(fastqc).version)
-            # Store the fastqc files needed for reporting
-            output_files.extend(["%s.html" % f for f in fastqcs])
-            output_files.extend([os.path.join(f,"summary.txt")
-                                 for f in fastqcs])
-            output_files.extend([os.path.join(f,"fastqc_data.txt")
-                                 for f in fastqcs])
-            # Fastqc plot images
-            for png in ("per_base_quality",):
-                output_files.extend([os.path.join(f,
-                                                  "Images",
-                                                  "%s.png" % png)
-                                     for f in fastqcs])
-        # Return collected information
-        if versions:
-            software = { 'fastqc': sorted(list(versions)) }
-        else:
-            software = {}
-        return AttributeDictionary(
-            name='fastqc',
-            software=software,
-            fastqs=sorted(list(fastqs)),
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_fastq_strand(self,files):
-        """
-        Collect information on FastqStrand outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'fastq_strand'
-        - software: dictionary of software and versions
-        - fastqs: list of associated Fastq names
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          files (list): list of file names to examine.
-        """
-        versions = set()
-        output_files = list()
-        fastqs = set()
-        tags = set()
-        # Look for fastq_strand outputs
-        fastq_strand = list(filter(lambda f:
-                                   f.endswith("_fastq_strand.txt"),
-                                   files))
-        logger.debug("fastq_strand: %s" % fastq_strand)
-        print("\t- %d fastq_strand files" % len(fastq_strand))
-        if fastq_strand:
-            tags.add("strandedness")
-            for f in fastq_strand:
-                fq = self.fastq_attrs(os.path.splitext(f)[0])
-                fastqs.add(
-                    os.path.basename(
-                        os.path.splitext(f)[0])[:-len("_fastq_strand")])
-                versions.add(Fastqstrand(f).version)
-            # Store the fastq_strand files
-            output_files.extend(fastq_strand)
-        # Return collected information
-        if versions:
-            software = { 'fastq_strand': sorted(list(versions)) }
-        else:
-            software = {}
-        return AttributeDictionary(
-            name='fastq_strand',
-            software=software,
-            fastqs=sorted(list(fastqs)),
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_seq_lengths(self,files):
-        """
-        Collect information on sequence length outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'sequence_lengths'
-        - software: dictionary of software and versions
-        - max_seqs: maximum number of sequences found in
-          a single Fastq
-        - min_seq_length: dictionary with minimum sequence
-          lengths for each read
-        - max_seq_length: dictionary with maximum sequence
-          lengths for each read
-        - reads: list of read IDs (e.g. 'r1', 'i2')
-        - fastqs: list of associated Fastq names
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          files (list): list of file names to examine.
-        """
-        output_files = list()
-        fastqs = set()
-        reads = set()
-        max_seqs = None
-        min_seq_length = {}
-        max_seq_length = {}
-        tags = set()
-        # Look for sequence length outputs
-        seq_lens = list(filter(lambda f:
-                               f.endswith("_seqlens.json"),
-                               files))
-        logger.debug("seq_lens: %s" % seq_lens)
-        print("\t- %d sequence length files" % len(seq_lens))
-        if seq_lens:
-            tags.add("sequence_lengths")
-            for f in seq_lens:
-                fq = self.fastq_attrs(os.path.splitext(f)[0])
-                read = '%s%s' % ('i' if fq.is_index_read else 'r',
-                                 fq.read_number)
-                fastqs.add(
-                    os.path.basename(
-                        os.path.splitext(f)[0])[:-len("_seqlens")])
-                reads.add(read)
-                seqlens_data = SeqLens(f)
-                try:
-                    # Try to extract the sequence lengths
-                    max_seqs = max(max_seqs,seqlens_data.nreads)
-                except Exception:
-                    if not max_seqs:
-                        max_seqs = seqlens_data.nreads
-                if not read in min_seq_length:
-                    min_seq_length[read] = seqlens_data.min_length
-                else:
-                    min_seq_length[read] = min(min_seq_length[read],
-                                               seqlens_data.min_length)
-                if not read in max_seq_length:
-                    max_seq_length[read] = seqlens_data.max_length
-                else:
-                    max_seq_length[read] = max(max_seq_length[read],
-                                               seqlens_data.max_length)
-            # Store the sequence length files
-            output_files.extend(seq_lens)
-        # Return collected information
-        return AttributeDictionary(
-            name='sequence_lengths',
-            software={},
-            max_seqs=max_seqs,
-            min_seq_length=min_seq_length,
-            max_seq_length=max_seq_length,
-            reads=sorted(list(reads)),
-            fastqs=sorted(list(fastqs)),
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_icell8(self,qc_dir):
-        """
-        Collect information on ICell8 reports
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'icell8'
-        - software: dictionary of software and versions
-        - fastqs: list of associated Fastq names
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          qc_dir (str): top-level directory to look under.
-        """
-        output_files = list()
-        tags = set()
-        # Look for ICELL8 outputs
-        icell8_top_dir = os.path.dirname(qc_dir)
-        print("Checking for ICELL8 reports in %s/stats" %
-              icell8_top_dir)
-        icell8_stats_xlsx = os.path.join(icell8_top_dir,
-                                         "stats",
-                                         "icell8_stats.xlsx")
-        if os.path.exists(icell8_stats_xlsx):
-            tags.add("icell8_stats")
-            output_files.append(icell8_stats_xlsx)
-        icell8_report_html = os.path.join(icell8_top_dir,
-                                          "icell8_processing.html")
-        if os.path.exists(icell8_report_html):
-            tags.add("icell8_report")
-            output_files.append(icell8_report_html)
-        # Return collected information
-        return AttributeDictionary(
-            name='icell8',
-            software={},
-            fastqs=[],
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_cellranger_count(self,qc_dir):
-        """
-        Collect information on Cellranger count outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'cellranger_count'
-        - software: dictionary of software and versions
-        - references: list of associated reference datasets
-        - fastqs: list of associated Fastq names
-        - samples: list of associated sample names
-        - pipelines: list of tuples defining 10x pipelines
-          in the form (name,version,reference)
-        - samples_by_pipeline: dictionary with lists of
-          sample names associated with each 10x pipeline
-          tuple
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          qc_dir (str): top-level directory to look under.
-        """
-        software = {}
-        output_files = list()
-        cellranger_samples = []
-        cellranger_references = set()
-        samples_by_pipeline = dict()
-        tags = set()
-        # Look for cellranger_count outputs
-        cellranger_count_dir = os.path.join(qc_dir,
-                                            "cellranger_count")
-        print("Checking for cellranger* count outputs under %s" %
-              cellranger_count_dir)
-        cellranger_versioned_samples = {}
-        if os.path.isdir(cellranger_count_dir):
-            cellranger_name = None
-            versions = set()
-            # Old-style (unversioned)
-            for d in filter(
-                    lambda f:
-                    os.path.isdir(os.path.join(cellranger_count_dir,f)),
-                    os.listdir(cellranger_count_dir)):
-                sample_dir = os.path.join(cellranger_count_dir,d)
-                try:
-                    cellranger = CellrangerCount(sample_dir)
-                    output_files.append(cellranger.web_summary)
-                    output_files.append(cellranger.metrics_csv)
-                    output_files.append(cellranger.cmdline_file)
-                    cellranger_samples.append(d)
-                    cellranger_name = cellranger.pipeline_name
-                    cellranger_references.add(cellranger.reference_data)
-                    # Store as version '?'
-                    ref = os.path.basename(cellranger.reference_data)
-                    if cellranger_name not in cellranger_versioned_samples:
-                        cellranger_versioned_samples[cellranger_name] = {}
-                    if '?' not in cellranger_versioned_samples[cellranger_name]:
-                        cellranger_versioned_samples[cellranger_name]['?'] = {}
-                    if ref not in \
-                       cellranger_versioned_samples[cellranger_name]['?']:
-                        cellranger_versioned_samples[cellranger_name]['?'][ref] = []
-                    cellranger_versioned_samples[cellranger_name]['?'][ref].append(d)
-                    versions.add('?')
-                except OSError:
-                    pass
-            if cellranger_samples:
-                tags.add("%s_count" % cellranger_name)
-            # New-style (versioned)
-            cellranger_name = None
-            for ver in filter(
-                    lambda f:
-                    os.path.isdir(os.path.join(cellranger_count_dir,f)),
-                    os.listdir(cellranger_count_dir)):
-                # Check putative version numbers
-                for ref in filter(
-                        lambda f:
-                        os.path.isdir(os.path.join(cellranger_count_dir,ver,f)),
-                        os.listdir(os.path.join(cellranger_count_dir,ver))):
-                    # Check putative reference dataset names
-                    samples = []
-                    for smpl in filter(
-                            lambda f:
-                            os.path.isdir(os.path.join(cellranger_count_dir,
-                                                       ver,ref,f)),
-                            os.listdir(os.path.join(cellranger_count_dir,
-                                                    ver,ref))):
-                        sample_dir = os.path.join(cellranger_count_dir,
-                                                  ver,ref,smpl)
-                        cellranger_name = None
-                        try:
-                            cellranger = CellrangerCount(sample_dir)
-                            output_files.append(cellranger.web_summary)
-                            output_files.append(cellranger.metrics_csv)
-                            output_files.append(cellranger.cmdline_file)
-                            samples.append(smpl)
-                            cellranger_name = cellranger.pipeline_name
-                            cellranger_references.add(
-                                cellranger.reference_data)
-                        except OSError:
-                            pass
-                    # Add outputs, samples and version
-                    if samples:
-                        tags.add("%s_count" % cellranger_name)
-                        if cellranger_name not in cellranger_versioned_samples:
-                            cellranger_versioned_samples[cellranger_name] = {}
-                        if ver not in cellranger_versioned_samples[cellranger_name]:
-                            cellranger_versioned_samples[cellranger_name][ver] = {}
-                        cellranger_versioned_samples[cellranger_name][ver][ref] = samples
-                        versions.add(ver)
-                        for smpl in cellranger_versioned_samples[cellranger_name][ver][ref]:
-                            if smpl not in cellranger_samples:
-                                cellranger_samples.append(smpl)
-            # Store cellranger versions
-            for cellranger_name in cellranger_versioned_samples:
-                software[cellranger_name] = sorted(list(cellranger_versioned_samples[cellranger_name].keys()))
-        # Store sample lists associated with pipeline,
-        # version and reference dataset
-        for name in cellranger_versioned_samples:
-            for version in cellranger_versioned_samples[name]:
-                for reference in cellranger_versioned_samples[name][version]:
-                    pipeline_key = (name,version,reference)
-                    samples_by_pipeline[pipeline_key] = \
-                        [s for s in
-                         cellranger_versioned_samples[name][version][reference]]
-        # Return collected information
-        return AttributeDictionary(
-            name='cellranger_count',
-            software=software,
-            references=sorted(list(cellranger_references)),
-            fastqs=[],
-            samples=cellranger_samples,
-            pipelines=sorted([p for p in samples_by_pipeline]),
-            samples_by_pipeline=samples_by_pipeline,
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_cellranger_multi(self,qc_dir):
-        """
-        Collect information on Cellranger multi outputs
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'cellranger_multi'
-        - software: dictionary of software and versions
-        - references: list of associated reference datasets
-        - fastqs: list of associated Fastq names
-        - multiplexed_samples: list of associated multiplexed
-          sample names
-        - pipelines: list of tuples defining 10x pipelines
-          in the form (name,version,reference)
-        - samples_by_pipeline: dictionary with lists of
-          multiplexed sample names associated with each 10x
-          pipeline tuple
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          qc_dir (str): top-level directory to look under.
-        """
-        software = {}
-        output_files = list()
-        multiplexed_samples = set()
-        cellranger_references = set()
-        samples_by_pipeline = dict()
-        tags = set()
-        # Look for cellranger multi outputs
-        cellranger_multi_dir = os.path.join(qc_dir,
-                                            "cellranger_multi")
-        print("Checking for cellranger multi outputs under %s" %
-              cellranger_multi_dir)
-        if os.path.isdir(cellranger_multi_dir):
-            cellranger_name = None
-            versions = set()
-            cellranger_multi_samples = {}
-            for ver in filter(
-                    lambda f:
-                    os.path.isdir(os.path.join(cellranger_multi_dir,f)),
-                    os.listdir(cellranger_multi_dir)):
-                cellranger_multi_samples[ver] = {}
-                for ref in filter(
-                        lambda f:
-                        os.path.isdir(os.path.join(cellranger_multi_dir,ver,f)),
-                        os.listdir(os.path.join(cellranger_multi_dir,ver))):
-                    # Check putative reference dataset names
-                    cellranger_multi_samples[ver][ref] = []
-                    cellranger_multi = CellrangerMulti(
-                        os.path.join(
-                            cellranger_multi_dir,
-                            ver,
-                            ref))
-                    for smpl in cellranger_multi.sample_names:
-                        cellranger_multi_samples[ver][ref].append(smpl)
-                        try:
-                            output_files.append(cellranger_multi.web_summary(smpl))
-                            output_files.append(cellranger_multi.metrics_csv(smpl))
-                            cellranger_name = cellranger_multi.pipeline_name
-                            if cellranger_name is None:
-                                cellranger_name = 'cellranger'
-                            cellranger_references.add(
-                                cellranger_multi.reference_data)
-                        except OSError:
-                            pass
-                    # Add outputs, samples and version
-                    if cellranger_multi_samples[ver][ref]:
-                        tags.add("cellranger_multi")
-                        versions.add(ver)
-                    for smpl in cellranger_multi_samples[ver][ref]:
-                        multiplexed_samples.add(smpl)
-            # Store sample lists associated with pipeline,
-            # version and reference dataset
-            for version in cellranger_multi_samples:
-                for reference in cellranger_multi_samples[version]:
-                    pipeline_key = (cellranger_name,version,reference)
-                    samples_by_pipeline[pipeline_key] = \
-                        [s for s in
-                         cellranger_multi_samples[version][reference]]
-            # Store cellranger versions
-            if cellranger_name and versions:
-                if cellranger_name not in software:
-                    software[cellranger_name] = list(versions)
-                else:
-                    for v in list(versions):
-                        if v not in software[cellranger_name]:
-                            software[cellranger_name].append(v)
-                software[cellranger_name] = sorted(software[cellranger_name])
-        # Return collected information
-        return AttributeDictionary(
-            name='cellranger_multi',
-            software=software,
-            references=sorted(list(cellranger_references)),
-            fastqs=[],
-            multiplexed_samples=sorted(list(multiplexed_samples)),
-            pipelines=sorted([p for p in samples_by_pipeline]),
-            samples_by_pipeline=samples_by_pipeline,
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_multiqc(self,qc_dir):
-        """
-        Collect information on MultQC reports
-
-        Returns an AttributeDictionary with the following
-        attributes:
-
-        - name: set to 'multiqc'
-        - software: dictionary of software and versions
-        - fastqs: list of associated Fastq names
-        - output_files: list of associated output files
-        - tags: list of associated output classes
-
-        Arguments:
-          qc_dir (str): top-level directory to look under.
-        """
-        version = None
-        output_files = list()
-        tags = set()
-        # Look for MultiQC report
-        multiqc_dir = os.path.dirname(qc_dir)
-        print("Checking for MultiQC report in %s" % multiqc_dir)
-        multiqc_report = os.path.join(multiqc_dir,
-                                      "multi%s_report.html"
-                                      % os.path.basename(qc_dir))
-        if os.path.isfile(multiqc_report):
-            tags.add("multiqc")
-            output_files.append(multiqc_report)
-            # Try to locate version from HTML file
-            # Look for line like e.g.
-            # <a href="http://multiqc.info" target="_blank">MultiQC v1.8</a>
-            with open(multiqc_report,'rt') as fp:
-                for line in fp:
-                    if line.strip().startswith("<a href=\"http://multiqc.info\" target=\"_blank\">MultiQC v"):
-                        try:
-                            version = line.strip().split()[3][1:-4]
-                        except Exception as ex:
-                            logger.warning("Failed to extract MultiQC version "
-                                           "from '%s': %s" % (line,ex))
-                        break
-        # Return collected information
-        if version:
-            software = { 'multiqc': [ version ] }
-        else:
-            software = {}
-        return AttributeDictionary(
-            name='multiqc',
-            software=software,
-            fastqs=[],
-            output_files=output_files,
-            tags=sorted(list(tags))
-        )
-
-    def _collect_qc_config_files(self,files):
-        """
-        Collect information on QC config files
-
-        Returns a list of QC configuration (if any) found
-        in the supplied list (with the leading paths
-        removed).
-
-        The following configuration files are checked for:
-
-        - fastq_strand.conf
-        - 10x_multi_config.csv
-        - libraries.<SAMPLE>.csv
-
-        Arguments:
-          files (list): list of file names to examine.
-        """
-        config_files = set()
-        # Strand and cellranger multi configs
-        for f in ("fastq_strand.conf",
-                  "10x_multi_config.csv",):
-            if os.path.join(self.qc_dir,f) in files:
-                config_files.add(f)
-        # Cellranger-arc configs ('libraries.SAMPLE.csv')
-        for f in list(filter(lambda f:
-                             f.endswith(".csv") and
-                             os.path.basename(f).startswith("libraries."),
-                             files)):
-            config_files.add(os.path.basename(f))
-        return sorted(list(config_files))
 
 class FastqSet:
     """
