+++ auto_process_ngs/test/qc/test_reporting.py
@@ -13,7 +13,6 @@ from auto_process_ngs.mockqc import MockQCOutputs
 from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.analysis import AnalysisSample
 from auto_process_ngs.qc.reporting import QCReporter
-from auto_process_ngs.qc.reporting import QCOutputs
 from auto_process_ngs.qc.reporting import FastqSet
 from auto_process_ngs.qc.reporting import verify
 from auto_process_ngs.qc.reporting import report
@@ -389,1358 +388,6 @@ PBB_CML2,CMO302,CML2
             'report.SE.PJB/qc/PJB2_S2_R1_001_rRNA_screen.txt')
         for f in expected:
             self.assertTrue(f in contents,"%s is missing from ZIP file" % f)
-
-class TestQCOutputs(unittest.TestCase):
-    def setUp(self):
-        # Temporary working dir (if needed)
-        self.wd = None
-    def tearDown(self):
-        # Remove temporary working dir
-        if not REMOVE_TEST_OUTPUTS:
-            return
-        if self.wd is not None and os.path.isdir(self.wd):
-            shutil.rmtree(self.wd)
-    def _make_working_dir(self):
-        # Create a temporary working directory
-        if self.wd is None:
-            self.wd = tempfile.mkdtemp(suffix='.test_QCOutputs')
-    def _make_qc_dir(self,qc_dir,fastq_names,
-                     screens=('model_organisms','other_organisms','rRNA',),
-                     cellranger_pipelines=('cellranger',),
-                     cellranger_samples=None,
-                     cellranger_multi_samples=None,
-                     include_fastqc=True,
-                     include_fastq_screen=True,
-                     include_strandedness=True,
-                     include_seqlens=True,
-                     include_multiqc=True,
-                     include_cellranger_count=False,
-                     include_cellranger_multi=False,
-                     legacy_screens=False,
-                     legacy_cellranger_outs=False):
-        # Create working directory and qc dir
-        self._make_working_dir()
-        qc_dir = os.path.join(self.wd,qc_dir)
-        os.mkdir(qc_dir)
-        # Populate with fake QC products
-        for fq in fastq_names:
-            # FastQC
-            if include_fastqc:
-                MockQCOutputs.fastqc_v0_11_2(fq,qc_dir)
-            # Fastq_screen
-            if include_fastq_screen:
-                for screen in screens:
-                    MockQCOutputs.fastq_screen_v0_9_2(
-                        fq,qc_dir,screen,
-                        legacy=legacy_screens)
-            # Strandedness
-            if include_strandedness:
-                MockQCOutputs.fastq_strand_v0_0_4(fq,qc_dir)
-            # Sequence lengths
-            if include_seqlens:
-                MockQCOutputs.seqlens(fq,qc_dir)
-        # Strandedness conf file
-        if include_strandedness:
-            with open(os.path.join(qc_dir,
-                                   "fastq_strand.conf"),'wt') as fp:
-                fp.write("Placeholder\n")
-        # MultiQC
-        if include_multiqc:
-            out_file = "multi%s_report.html" % os.path.basename(qc_dir)
-            MockQCOutputs.multiqc(self.wd,
-                                  multiqc_html=out_file,
-                                  version="1.8")
-        # Cellranger count
-        if include_cellranger_count:
-            for cellranger in cellranger_pipelines:
-                # Set defaults
-                if cellranger == "cellranger":
-                    version = "6.1.2"
-                    refdata = "/data/refdata-cellranger-2020-A"
-                elif cellranger == "cellranger-atac":
-                    version = "2.0.0"
-                    refdata = "/data/refdata-cellranger-atac-2020-A"
-                elif cellranger == "cellranger-arc":
-                    version = "2.0.0"
-                    refdata = "/data/refdata-cellranger-arc-2020-A"
-                # Set top-level output dir
-                if not legacy_cellranger_outs:
-                    count_dir = os.path.join("cellranger_count",
-                                             version,
-                                             os.path.basename(refdata))
-                else:
-                    count_dir = "cellranger_count"
-                # Make pipeline outputs
-                for sample in cellranger_samples:
-                    MockQCOutputs.cellranger_count(
-                        sample,
-                        qc_dir,
-                        cellranger=cellranger,
-                        version=version,
-                        reference_data_path=refdata,
-                        prefix=count_dir)
-                    if cellranger == "cellranger-arc":
-                        multiome_config = os.path.join(qc_dir,
-                                                       "libraries.%s.csv" %
-                                                       sample)
-                        with open(multiome_config,'wt') as fp:
-                            fp.write("Placeholder\n")
-        # Cellranger multi
-        if include_cellranger_multi:
-            # Make cellranger multi config.csv file
-            multi_config = os.path.join(qc_dir,"10x_multi_config.csv")
-            with open(multi_config,'wt') as fp:
-                fastq_dir = os.path.join(self.wd,
-                                         "PJB",
-                                         "fastqs")
-                fp.write("""[gene-expression]
-reference,/data/refdata-cellranger-2020-A
-
-[libraries]
-fastq_id,fastqs,lanes,physical_library_id,feature_types,subsample_rate
-PJB1_GEX,%s,any,PJB1,gene expression,
-PJB2_MC,%s,any,PJB2,Multiplexing Capture,
-
-[samples]
-sample_id,cmo_ids,description
-PJB_CML1,CMO301,CML1
-PJB_CML2,CMO302,CML2
-""" % (fastq_dir,fastq_dir))
-            # Set top-level output dir
-            multi_dir = os.path.join("cellranger_multi",
-                                     "6.1.2",
-                                     "refdata-cellranger-2020-A")
-            # Make outputs
-            MockQCOutputs.cellranger_multi(cellranger_multi_samples,
-                                           qc_dir,
-                                           config_csv=multi_config,
-                                           prefix=multi_dir)
-        return qc_dir
-
-    def test_qcoutputs_single_end(self):
-        """
-	QCOutputs: single-end data
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'multiqc',
-                          'screens_r1',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_single_end_no_fastqc(self):
-        """
-	QCOutputs: single-end data (no FastQC)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   include_fastqc=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['multiqc',
-                          'screens_r1',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_single_end_no_screens(self):
-        """
-	QCOutputs: single-end data (no screens)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   include_fastq_screen=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'multiqc',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,[])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_single_end_no_seqlens(self):
-        """
-	QCOutputs: single-end data (no sequence lengths)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   include_seqlens=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'multiqc',
-                          'screens_r1',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,None)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,None)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,None)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         [])
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_single_end_no_standedness(self):
-        """
-	QCOutputs: single-end data (no strandedness)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   include_strandedness=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'multiqc',
-                          'screens_r1',
-                          'sequence_lengths'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,[])
-
-    def test_qcoutputs_single_end_no_multiqc(self):
-        """
-	QCOutputs: single-end data (no MultiQC)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   include_multiqc=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'screens_r1',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_single_end_legacy_screen_naming(self):
-        """
-	QCOutputs: single-end data (legacy screen naming)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB2_S2_R1_001',
-                                   ),
-                                   legacy_screens=True)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'multiqc',
-                          'screens_r1',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB2_S2_R1_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end(self):
-        """
-	QCOutputs: paired-end data
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end_no_fastqc(self):
-        """
-	QCOutputs: paired-end data (no FastQC)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_fastqc=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end_no_screens(self):
-        """
-	QCOutputs: paired-end data (no screens)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_fastq_screen=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,[])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end_no_seqlens(self):
-        """
-	QCOutputs: paired-end data (no sequence lengths)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_seqlens=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,None)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,None)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,None)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),[])
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end_no_strandedness(self):
-        """
-	QCOutputs: paired-end data (no strandedness)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_strandedness=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,[])
-
-    def test_qcoutputs_paired_end_no_multiqc(self):
-        """
-	QCOutputs: paired-end data (no MultiQC)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_multiqc=False)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_paired_end_legacy_screen_naming(self):
-        """
-	QCOutputs: paired-end data (legacy screen naming)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   legacy_screens=True)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_10x_cellranger_count(self):
-        """
-        QCOutputs: 10xGenomics data with cellranger 'count'
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   cellranger_pipelines=('cellranger',),
-                                   cellranger_samples=(
-                                       'PJB1',
-                                       'PJB2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger_count',
-                          'fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger': [ '6.1.2' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_10x_cellranger_count_legacy(self):
-        """
-        QCOutputs: 10xGenomics data with cellranger 'count' (legacy format)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   cellranger_pipelines=('cellranger',),
-                                   cellranger_samples=(
-                                       'PJB1',
-                                       'PJB2',
-                                   ),
-                                   legacy_cellranger_outs=True)
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger_count',
-                          'fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger': [ '?' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_10x_cellranger_atac_count(self):
-        """
-        QCOutputs: 10xGenomics data with cellranger-atac 'count'
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R3_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R3_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   cellranger_pipelines=('cellranger-atac',),
-                                   cellranger_samples=(
-                                       'PJB1',
-                                       'PJB2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger-atac_count',
-                          'fastqc_r1',
-                          'fastqc_r3',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r3',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R3_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R3_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-atac-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r3'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger-atac': [ '2.0.0' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r3'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r3'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r3'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_10x_multiome_gex(self):
-        """
-        QCOutputs: 10xGenomics multiome data (GEX component)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   cellranger_pipelines=('cellranger',
-                                                         'cellranger-arc',),
-                                   cellranger_samples=(
-                                       'PJB1',
-                                       'PJB2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger-arc_count',
-                          'cellranger_count',
-                          'fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-2020-A',
-                          '/data/refdata-cellranger-arc-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger': [ '6.1.2' ],
-                           'cellranger-arc': [ '2.0.0' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf',
-                          'libraries.PJB1.csv',
-                          'libraries.PJB2.csv'])
-
-    def test_qcoutputs_10x_multiome_atac(self):
-        """
-        QCOutputs: 10xGenomics multiome data (ATAC component)
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R3_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R3_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   cellranger_pipelines=('cellranger-arc',
-                                                         'cellranger-atac',),
-                                   cellranger_samples=(
-                                       'PJB1',
-                                       'PJB2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger-arc_count',
-                          'cellranger-atac_count',
-                          'fastqc_r1',
-                          'fastqc_r3',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r3',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R3_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R3_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-arc-2020-A',
-                          '/data/refdata-cellranger-atac-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r3'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger-arc': [ '2.0.0' ],
-                           'cellranger-atac': [ '2.0.0' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r3'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r3'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r3'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf',
-                          'libraries.PJB1.csv',
-                          'libraries.PJB2.csv'])
-
-    def test_qcoutputs_10x_cellranger_multi(self):
-        """
-        QCOutputs: 10xGenomics CellPlex data with 'cellranger multi'
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_GEX_S1_R1_001',
-                                       'PJB1_GEX_S1_R2_001',
-                                       'PJB2_MC_S2_R1_001',
-                                       'PJB2_MC_S2_R2_001',
-                                   ),
-                                   include_cellranger_multi=True,
-                                   cellranger_pipelines=('cellranger',),
-                                   cellranger_samples=(
-                                       'PJB1_GEX',
-                                       'PJB2_MC',
-                                   ),
-                                   cellranger_multi_samples=(
-                                       'PJB_CML1',
-                                       'PJB_CML2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger_multi',
-                          'fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_GEX_S1_R1_001',
-                          'PJB1_GEX_S1_R2_001',
-                          'PJB2_MC_S2_R1_001',
-                          'PJB2_MC_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1_GEX','PJB2_MC'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,
-                         ['PJB_CML1','PJB_CML2'])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger': [ '6.1.2' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['10x_multi_config.csv',
-                          'fastq_strand.conf'])
-
-    def test_qcoutputs_10x_cellranger_multi_and_count(self):
-        """
-        QCOutputs: 10xGenomics CellPlex data with 'cellranger multi' and 'count'
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_GEX_S1_R1_001',
-                                       'PJB1_GEX_S1_R2_001',
-                                       'PJB2_MC_S2_R1_001',
-                                       'PJB2_MC_S2_R2_001',
-                                   ),
-                                   include_cellranger_count=True,
-                                   include_cellranger_multi=True,
-                                   cellranger_pipelines=('cellranger',),
-                                   cellranger_samples=(
-                                       'PJB1_GEX',
-                                       'PJB2_MC',
-                                   ),
-                                   cellranger_multi_samples=(
-                                       'PJB_CML1',
-                                       'PJB_CML2',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['cellranger_count',
-                          'cellranger_multi',
-                          'fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_GEX_S1_R1_001',
-                          'PJB1_GEX_S1_R2_001',
-                          'PJB2_MC_S2_R1_001',
-                          'PJB2_MC_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1_GEX','PJB2_MC'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,
-                         ['/data/refdata-cellranger-2020-A'])
-        self.assertEqual(qc_outputs.multiplexed_samples,
-                         ['PJB_CML1','PJB_CML2'])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'cellranger': [ '6.1.2' ],
-                           'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['10x_multi_config.csv',
-                          'fastq_strand.conf'])
-
-    def test_qcoutputs_with_non_standard_qc_dir(self):
-        """
-        QCOutputs: non-standard QC directory
-        """
-        qc_dir = self._make_qc_dir('my_qc_outs',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001',
-                                       'PJB1_S1_R2_001',
-                                       'PJB2_S2_R1_001',
-                                       'PJB2_S2_R2_001',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001',
-                          'PJB1_S1_R2_001',
-                          'PJB2_S2_R1_001',
-                          'PJB2_S2_R2_001'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
-
-    def test_qcoutputs_with_non_canonical_fastq_names(self):
-        """
-        QCOutputs: non-canonical fastq names
-        """
-        qc_dir = self._make_qc_dir('qc',
-                                   fastq_names=(
-                                       'PJB1_S1_R1_001_paired',
-                                       'PJB1_S1_R2_001_paired',
-                                       'PJB2_S2_R1_001_paired',
-                                       'PJB2_S2_R2_001_paired',
-                                   ))
-        qc_outputs = QCOutputs(qc_dir)
-        self.assertEqual(qc_outputs.outputs,
-                         ['fastqc_r1',
-                          'fastqc_r2',
-                          'multiqc',
-                          'screens_r1',
-                          'screens_r2',
-                          'sequence_lengths',
-                          'strandedness'])
-        self.assertEqual(qc_outputs.fastqs,
-                         ['PJB1_S1_R1_001_paired',
-                          'PJB1_S1_R2_001_paired',
-                          'PJB2_S2_R1_001_paired',
-                          'PJB2_S2_R2_001_paired'])
-        self.assertEqual(qc_outputs.samples,
-                         ['PJB1','PJB2'])
-        self.assertEqual(qc_outputs.fastq_screens,
-                         ['model_organisms',
-                          'other_organisms',
-                          'rRNA'])
-        self.assertEqual(qc_outputs.cellranger_references,[])
-        self.assertEqual(qc_outputs.multiplexed_samples,[])
-        self.assertEqual(qc_outputs.reads,['r1','r2'])
-        self.assertEqual(qc_outputs.software,
-                         { 'fastqc': [ '0.11.3' ],
-                           'fastq_screen': [ '0.9.2' ],
-                           'fastq_strand': [ '0.0.4' ],
-                           'multiqc': [ '1.8' ],
-                         })
-        self.assertEqual(qc_outputs.stats.max_seqs,37285443)
-        self.assertEqual(qc_outputs.stats.min_sequence_length,65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length,76)
-        self.assertEqual(sorted(
-            list(qc_outputs.stats.min_sequence_length_read.keys())),
-                         ['r1','r2'])
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r1'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r1'],76)
-        self.assertEqual(qc_outputs.stats.min_sequence_length_read['r2'],65)
-        self.assertEqual(qc_outputs.stats.max_sequence_length_read['r2'],76)
-        self.assertEqual(qc_outputs.config_files,
-                         ['fastq_strand.conf'])
 
 class TestFastqSet(unittest.TestCase):
     def test_fastqset_PE(self):
